refactor: [DHIS2-10250] rename EnrollmentEvent to EnrollmentEditEvent
