+++ cypress/integration/EnrollmentEditEventPage/EnrollmentEditEventPageNavigation.feature
@@ -1,0 +1,6 @@
+Feature: User interacts with Enrollment event page
+
+  Scenario: The user can land on the enrollment event page.
+    Given you land on the enrollment event page by having typed #/enrollmentEventEdit?programId=IpHINAT79UW&orgUnitId=DiszpKrYNg&teiId=EaOyKGOIGRp&enrollmentId=IMEGtAd4DOj&stageId=ZzYYXq4fJie&eventId=vDGrlVeyE2Q
+    Then you see the following Enrollment: View Event
+    And you see the following Baby Postnatal
