+++ cypress/integration/EnrollmentEditEventPage/EnrollmentEditEventPageNavigation/index.js
