--- cypress/integration/EnrollmentEventPage/EnrollmentEventPageNavigation.feature
@@ -1,6 +1,0 @@
-Feature: User interacts with Enrollment event page
-
-  Scenario: The user can land on the enrollment event page.
-    Given you land on the enrollment event page by having typed #/enrollment/event/edit?programId=IpHINAT79UW&orgUnitId=DiszpKrYNg&teiId=EaOyKGOIGRp&enrollmentId=IMEGtAd4DOj&stageId=ZzYYXq4fJie&eventId=vDGrlVeyE2Q
-    Then you see the following Enrollment: View Event
-    And you see the following Baby Postnatal
