+++ src/components/App/AppPages.component.js
@@ -7,7 +7,7 @@ import { SearchPage } from 'capture-core/components/Pages/Search';
 import { NewPage } from 'capture-core/components/Pages/New';
 import { EnrollmentPage } from 'capture-core/components/Pages/Enrollment';
 import { StageEventListPage } from 'capture-core/components/Pages/StageEvent';
-import { EnrollmentEventPage } from 'capture-core/components/Pages/EnrollmentEvent';
+import { EnrollmentEditEventPage } from 'capture-core/components/Pages/EnrollmentEditEvent';
 import { EnrollmentAddEventPage } from 'capture-core/components/Pages/EnrollmentAddEvent';
 
 export const AppPages = () => (
@@ -16,7 +16,7 @@ export const AppPages = () => (
         <Route path="/search" component={SearchPage} />
         <Route path="/new" component={NewPage} />
         <Route path="/enrollment/stageEvents" component={StageEventListPage} />
-        <Route path="/enrollment/event/edit" component={EnrollmentEventPage} />
+        <Route path="/enrollmentEventEdit" component={EnrollmentEditEventPage} />
         <Route path="/enrollmentEventNew" component={EnrollmentAddEventPage} />
         <Route path="/enrollment" component={EnrollmentPage} />
         <Route path="/:keys" component={MainPage} />
