+++ src/core_modules/capture-core/components/Pages/EnrollmentEditEvent/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { EnrollmentEditEventPage } from './EnrollmentEditEventPage.container';
