--- src/core_modules/capture-core/components/Pages/EnrollmentEvent/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { EnrollmentEventPage } from './EnrollmentEventPage.container';
