+++ src/core_modules/capture-core/components/WidgetEventEdit/WidgetEventEdit.container.js
@@ -6,7 +6,7 @@ import { Widget } from '../Widget';
 import type { Props } from './widgetEventEdit.types';
 import { EditEventDataEntry } from './EditEventDataEntry/';
 import { ViewEventDataEntry } from './ViewEventDataEntry/';
-import { pageMode } from '../Pages/EnrollmentEvent/EnrollmentEventPage.const';
+import { pageMode } from '../Pages/EnrollmentEditEvent/EnrollmentEditEventPage.const';
 import { NonBundledDhis2Icon } from '../NonBundledDhis2Icon';
 
 const styles = {
